macro_agents/pyproject.toml
[project]
name = "macro_agents"
requires-python = ">=3.10.19,<3.14"
version = "0.1.0"
dependencies = [
    "dagster>=1.11.4,<2.0.0",  # Add upper bound
    "dagster-cloud",
    "dagster-duckdb",
    "polars>=0.20.0,<1.0.0",  # Pin polars
    "dagster-dbt",
    "requests>=2.31.0,<3.0.0",
    "dbt-core>=1.10.0,<2.0.0",  # Pin dbt-core
    "google-api-core>=2.0.0,<3.0.0",
    "google-auth>=2.0.0,<3.0.0",
    "google-cloud-bigquery>=3.0.0,<4.0.0",
    "dbt-duckdb>=1.10.0,<2.0.0",
    "dagster-shared>=1.11.4,<2.0.0",
    "dagster-pipes>=1.11.4,<2.0.0",
    "dagster-dg-cli>=1.11.4,<2.0.0",
    "dagster-webserver>=1.11.4,<2.0.0",
    "pyarrow>=21.0.0,<22.0.0",  # Pin pyarrow
    "beautifulsoup4>=4.12.0,<5.0.0",
    "lxml>=5.0.0,<6.0.0",
    # Move dev tools to optional-dependencies
    "tenacity>=9.1.2,<10.0.0",
    "dspy>=2.6.27,<3.0.0",
    "dagster-sling",
    "google-cloud-storage>=2.10.0,<3.0.0",
    "openai>=1.0.0,<2.0.0",
    "anthropic>=0.18.0,<1.0.0",
    "google-generativeai>=0.3.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "dagster-webserver", 
    "pytest",
    "pytest-cov",
    "pytest-xdist",
    "mypy",
    "ruff>=0.12.7,<1.0.0",  # Move ruff here
    "sqlfluff>=3.4.2,<4.0.0",  # Move sqlfluff here
    "sqlfluff-templater-dbt>=3.4.2,<4.0.0",  # Move here
]
 
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/macro_agents"]
include = [
    "dbt_project/**/*",
]

[tool.dg]
directory_type = "project"

[tool.dg.project]
root_module = "macro_agents"

[tool.pytest.ini_options]
markers = [
    "essential: Essential tests that must pass in CI",
    "skip_ci: Tests to skip in CI (brittle or non-essential)",
]
filterwarnings = [
    "ignore::pydantic.warnings.PydanticDeprecatedSince20:",
    "ignore::DeprecationWarning:pydantic._internal._config",
]


